atomic-forge/tools/bocha_search/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tool"]

[project]
name = "bocha-search"
version = "1.0.0"
description = "A tool for performing searches using BoCha"
readme = "README.md"
authors = [
    { name = "Jintao Wei", email = "2797154930@qq.com" }
]
requires-python = ">=3.12"
dependencies = [
    "atomic-agents",
    "pydantic>=2.10.3,<3.0.0",
    "aiohttp>=3.9.0,<4.0.0",
]

[dependency-groups]
dev = [
    "coverage>=7.6.1,<8.0.0",
    "pytest>=8.3.3,<9.0.0",
    "pytest-asyncio>=0.23.5,<1.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "rich>=13.7.0,<14.0.0",
]

[tool.uv.sources]
atomic-agents = { workspace = true }
